Round up min gas ascent time
When calculating ascent time by hand, people use minutes and round up
fractional values to build in conservativism. We calculate with seconds
for precision in intermediate steps, but after calculating total ascent
time, round it up to the nearest minute.

This also ensures consistency in the min gas display; since the display
shows time values in minutes, doing this means we use the exact time
values that are displayed in our math. If we don't do this then the gas
values can change without the displayed minute values changing which,
while still accurate, is likely to confuse users.