Generate complete segment table for all depths
Rather than limiting the segment table to end at the planned bottom depth,
generate all possible segments in the supported depth range of the program and
display the table offset to the planned depth.

When doing this for meters we generate more segment table entries than can fit
in g1 memory (200 apparently) and hit OOM. To avoid this, limit max program
depth to 100m, still reasonable.

Also flip the scroll indicators so + is up and - is down. Makes more sense in
segment tables.